refactor: ♻️ projects tabs
